Create Final Code Submission
This was the final code I developed to autonomously search for an infrared beacon that pulsed every 100ms. The robot would turn on the spot until it detected high IR, then move towards the source of the IR, adjusting to stay on course as it advanced. When the robot finally reached the beacon, it connected a trailing cable, and simultaneously disconnected itself from the cable.