Create IR sensor Milestone
For this milestone, our group decided to build two BJT circuits using phototransistors to detect whether our robot was facing the beacon.
I was in charge of writing code for each of the four tests associated with this milestone to ensure our robot responded correctly to the IR stimulus, or for part D, sonar stimulus. The tasks were as follows:

a)Turn  your  robot  towards  the  target  object.	This  function  should  operate  properly when your robot is within 1 metre of the target.
b) Signal when your robot is properly facing the target object for approach but is  too  far  away  for  cable  connection.  This  function  should  be  able  to  function  properly when your robot is within 1 metre of the target.
 c) Signal  when  your  robot  is  at  the  proper  distance  and  orientation  for  cable  connection.
d) Signal when your robot has made contact with a wall when turning or moving  straight.