Create Mobility Milestone
This was the first milestone that involved coding, and the final result is perhaps a little less clean than the later versions, as I continued to learn and improve throughout the process. For this milestone, our group had to build a strong chassis for our robot, as well as select wheels and gear ratios were thought would be appropriate and implement them. My role was to design code which at the touch of a button cause the following effects in separate tests:
a) Cause the robot to drive forward one metre.
b) Cause the robot to turn 90 degrees.
c) Activate the cable release mechanism.